Lightbulb: Removed the reg key installation post_install
While installing the program you get a pop-up from the registry editor to ask permission if you don't run Scoop with admin rights. This can be seen as mallicious and it only need to be run once if the gamma doesn't change under 6600K. I've added the latter part in the notes. It's still long though.